Merlions Full First Class Configuration
http://theomegahangar.flymerlion.org/aircraft.html
<--------------------------------README----------------------------------->
This package includes the CRJ-700 Full First Class configuration aircraft.
The systems, flightdeck, fdm and other aircraft related characterestics
have not been changed. It includes merlion's CRJ700 FFC configuration. That
is 12 rows, of 2 seats in a row except the last one. This provides the VA
pax more comfort and a complete aircraft on First Class for highly popular
routes.

To install, Merge the CRJ700-family folder here into the CRJ700-family
folder in your fgdata folder.

UPDATE! - Now, the cabin design changes with livery! So don't forget to make
a cabin design too when you make a livery!

To run, select it from FGRun or start with the option '--aircraft=CRJ700FFC'

Contact 'contact.merlion@gmail.com' for more information.
<------------------------------------------------------------------------->

Note - This was in the pilotshop but now, a free download is available as:
1. It is required for pilots who wish to fly these FFC routes.
2. We have not decided on someone to manage the pilot shop.

If you'd like to apply, please contact the adminsitrators.

Cheers!
-Narendran